broke out query to procedural module
