+++ tests/test_space_object.py
@@ -4,7 +4,7 @@ sys.path.append("..")
 from mock import sbs
 
 from sbs_utils.spaceobject import SpaceObject, TickType
-from sbs_utils.query import role, linked_to, closest_list, closest, broad_test, to_py_object_list
+from sbs_utils.sbs_utils.procedural.query import role, linked_to, closest_list, closest, broad_test, to_py_object_list
 from sbs_utils.objects import Npc, Terrain, PlayerShip
 from sbs_utils.gridobject import GridObject
 from sbs_utils.helpers import FrameContext, Context
